Numerous small changes to prepare for a new alpha release: 1.1a2.
Most of these are compiler warnings that have been squashed at the
request of Rocky.  Some of the more significant changes appear below:

- Commented out the GML code in the cartography module because it
  appears to be half finished.

- Fixed a bug in the Linescan model that was causing the linetimes to
  be truncated to integer values.

- Fixed a bug in OrthoImageView -- pixel indexing is now integer
  based, not floating point based.

+++ src/vw/Cartography/OrthoImageView.h
@@ -81,7 +81,7 @@ namespace cartography {
     /// compute a 3D point corresponding to this location in the DTM.
     /// This point is then "imaged" by the camera model and the
     /// resulting pixel location is returned from the camera image.
-    inline result_type operator()( double i, double j, int p=0 ) const { 
+    inline result_type operator()( int32 i, int32 j, int p=0 ) const { 
 
       // We need to convert the georefernced positions into a
       // cartesian coordinate system so that they can be imaged by the
